Read in a file from a GitLab repository

Downloads the file under the specified path from a GitLab repository via GitLab's v4 RESTful API and returns its content.

Usage

gl_file_raw(
  path,
  id_proj,
  ref = NULL,
  base_url = pal::pkg_config_val("gitlab_base_url"),
  token = pal::pkg_config_val("gitlab_token"),
  max_tries = 3L
)

Arguments

path

File path, relative to the repository root. A character scalar.

id_proj

GitLab project identifier. An integerish number.

ref

Git revision expression matching the desired Git tree object, e.g. a ref name (branch, tag, etc.), a commit identifier, or another symbolic reference like "HEAD~10". A character scalar. Note that the GitLab API doesn't support every type of revision expression.

base_url

Base URL to the GitLab v4 RESTful API root.

token

GitLab access token used for authentication

max_tries

Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.

Value

A character scalar if path refers to a text file, otherwise a raw vector.

See also

Examples

yay::gl_file_raw(path = "DESCRIPTION",
                 id_proj = 24995012) |>
  cat()
#> Package: yay
#> Title: Delightful Convenience Functions
#> Version: 0.0.9013
#> Authors@R: 
#>     person("Salim", "Brüggemann", , "salim-b@pm.me", role = c("aut", "cre"),
#>            comment = c(ORCID = "0000-0002-5329-5987"))
#> Description: A delightful mix of convenience functions designed to put a
#>     smile on your face during typical R workflows.
#> License: AGPL (>= 3)
#> URL: https://yay.rpkg.dev, https://gitlab.com/rpkg.dev/yay
#> BugReports: https://gitlab.com/rpkg.dev/yay/issues
#> Depends: 
#>     R (>= 4.4)
#> Imports: 
#>     checkmate (>= 2.0.0),
#>     cli (>= 3.0.0),
#>     dplyr (>= 1.1.1),
#>     fs (>= 1.5.0),
#>     glue (>= 1.4.2),
#>     httr2 (>= 1.0.5),
#>     magrittr (>= 2.0.1),
#>     pal (>= 0.0.9096),
#>     purrr (>= 0.3.4),
#>     rlang (>= 0.4.10),
#>     stringr (>= 1.5.0),
#>     tibble (>= 3.1.0),
#>     tidyr (>= 1.1.3)
#> Suggests: 
#>     brio (>= 1.1.1),
#>     daff (>= 0.3.8),
#>     desc (>= 1.3.0),
#>     gert (>= 1.2.0),
#>     gh (>= 1.4.0),
#>     jsonlite (>= 1.8.7),
#>     pkgdown (>= 2.0.9.9000),
#>     pkgdown.tpl (>= 0.0.9000),
#>     pkgsnip (>= 0.0.9037),
#>     readr (>= 1.4.0),
#>     roxygen2 (>= 7.1.1),
#>     testthat (>= 3.0.0),
#>     withr (>= 2.4.1),
#>     writexl (>= 1.3.1),
#>     xopen (>= 1.0.0),
#>     yesno (>= 0.1.2),
#>     zoterro (>= 0.1.9001)
#> Remotes:
#>     github::edwindj/daff,
#>     github::salim-b/zoterro@interim-merge,
#>     gitlab::rpkg.dev/pal,
#>     gitlab::rpkg.dev/pkgdown.tpl,
#>     gitlab::rpkg.dev/pkgsnip
#> Config/testthat/edition: 3
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.3.2.9000