Write file to GitLab repository

Uploads a file to a GitLab repository via GitLab's v4 RESTful API.

Usage

gl_file_write(
  content,
  path,
  id_proj,
  start_branch = gl_branch_default(id_proj = id_proj, base_url = base_url, token = token,
    max_tries = max_tries),
  branch = start_branch,
  from_file = FALSE,
  overwrite = TRUE,
  force = FALSE,
  commit_message = "auto: write file via yay R pkg",
  author_email = NULL,
  author_name = NULL,
  execute_filemode = FALSE,
  last_commit_id = gl_file_meta(path = path, attribute = "last_commit_id", id_proj =
    id_proj, ref = branch, base_url = base_url, token = token, max_tries = max_tries),
  base_url = pal::pkg_config_val("gitlab_base_url"),
  token = pal::pkg_config_val("gitlab_token"),
  max_tries = 3L,
  quiet = FALSE
)

Arguments

content

File content, as a character scalar for text files, or a raw vector for binary files. Or the path to a local file as a character scalar if from_file = TRUE.

path

File path, relative to the repository root. A character scalar.

id_proj

GitLab project identifier. An integerish number.

start_branch

Name of the base branch to create branch from if it doesn't already exist.

branch

Git branch name to upload the file to.

from_file

Whether or not content indicates the path to a local file instead of the actual file content.

overwrite

Whether or not to overwrite an already existing file. Unless force = TRUE, the file is only overwritten if it is not identical to content.

force

Whether or not to overwrite an already existing file even if it is identical to content.

commit_message

Git commit message.

author_email

Git commit author's email address. NULL means to default to the token owner.

author_name

Git commit author's name. NULL means to default to the token owner.

execute_filemode

Whether or not to mark the uploaded file as executable,

last_commit_id

Last known file commit identifier. If provided (and valid), GitLab won't write to the file if a commit made after last_commit_id has modified the file. Intended to avoid concurrency issues (e.g. from CI pipelines). NULL means to skip the check.

base_url

Base URL to the GitLab v4 RESTful API root.

token

GitLab access token used for authentication

max_tries

Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.

quiet

Whether or not to suppress printing status output from internal processing.

Value

path, invisibly.

Details

To save up to two additional HEAD requests, set start_branch and last_commit_id explicitly. Setting last_commit_id = NULL is fine as long as file writes are part of a controlled, single-user process where concurrency is not a concern.

See also