Regular expression patterns and replacements for file normalization

Regular expression patterns and replacements for file normalization

Usage

regex_file_normalization

Format

A tibble.

See also

Examples

# unnest the pattern column
tidyr::unnest_longer(data = yay::regex_file_normalization,
                     col = pattern)
#> # A tibble: 3 × 5
#>   id                                     category    purpose                                                                                     pattern repla…¹
#>   <chr>                                  <chr>       <chr>                                                                                       <chr>   <chr>  
#> 1 terminating_newline                    posix       "ensure file ends with the [POSIX-standard newline control character `LF`](https://en.wiki… "(.)(?… "\\1\n"
#> 2 go_template_opening_double_brace_space go_template "ensure there's a space after opening double braces in Go templates"                        "\\{\\… "{{ "  
#> 3 go_template_closing_double_brace_space go_template "ensure there's a space before closing double braces in Go templates"                       "([^-%… "\\1 }…
#> # … with abbreviated variable name ¹​replacement