yay: Delightful Convenience Functions

A delightful mix of convenience functions designed to put a smile on your face during typical R workflows.

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)