Deploy static website to local Git folder

Deploy static website to local Git folder

deploy_static_site(
  from_path,
  to_path,
  clean_to_path = TRUE,
  commit_msg = "auto-deploy static website"
)

Arguments

from_path
Path to the directory containing the static website files that are to be deployed. A character scalar.

to_path
Path to the Git (sub)folder to which the static website files are to be deployed. A character scalar.

clean_to_path
Whether or not to completely wipe to_path before deploying the new website files. Setting this to TRUE ensures there are no obsolete files left over from previous deployments.

commit_msg
The Git commit message used for the deployment. A character scalar.

Value

A vector of paths to the deployed files/folders, invisibly.