yay: Delightful Convenience Functions

A delightful mix of convenience functions designed to put a smile on your face during typical R workflows.

Package configuration

Some of yay's functionality is controlled via package-specific global configuration which can either be set via R options or environment variables (the former take precedence). This configuration includes:

Description	R option	Environment variable	Default value

Netlify personal access token used for authentication in the netlify_*() functions.	yay.netlify_token	R_YAY_NETLIFY_TOKEN

Porkbun API key used for authentication in the porkbun_*() functions.	yay.porkbun_api_key	R_YAY_PORKBUN_API_KEY

Porkbun secret API key used for authentication in the porkbun_*() functions.	yay.porkbun_secret_api_key	R_YAY_PORKBUN_SECRET_API_KEY

See also

Author

Maintainer: Salim Brüggemann salim@rpkg.dev (ORCID)